Bring specific Member block settings into the Block Editor
In order to be consistent with the `show_avatars` option, we need to bring this site owner preference into the Block Editor settings.
As the Member block will not be the only one to need to bring some options, the idea is to gather all these settings into a `bp` key of the editor settings object.
Bringing these options will also help us to build the block settings sidebar.

See #3