Start building a block to embed a BP Activity
The regular WP Embed block isn’t rendering the activity preview the right way and I believe it’s best to have our own.
One thing we’re doing differently comparing to the WP Embed block is using the `scripts` property of the Sandbox component to let `wp-embed.js` hide the blockquote and display the iframe containing the activity preview.