First step at building the BP Members Block
The idea of this block is to let the post/page author to select the members he wants to add to the members list.

The BP Autocompleter make it easy to select members one by one. If we are using a dynamic block, we can't use the ServerSideRender component to output the block.
We need to deal with it using JavaScript so that it's possible to remove a member from the list.