Bring back Block Components & Member/Group blocks
It's more convenient to maintain all blocks from this repository.
To do so we added some hooks to use all BuddyPress blocks from this plugin instead of the BuddyPress ones.
We are also using a namespace so that it will be easier to merge into BP Core blocks/functions when they will be ready to.